story specの追加と通すための準備
