+++ backend/config/routes.rb
@@ -1,4 +1,7 @@
 Rails.application.routes.draw do
-  resources :stories
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+  namespace 'api' do
+    namespace 'v1' do
+      resources :stories
+    end
+  end
 end
