+++ backend/spec/factories/stories.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :story, class: Story do
+    title { 'title' }
+  end
+end
